Change SP init in fancy-register-test
I don't know why yet, but when I initialize the stack pointer to
a value with the three lowest-order bits high (i.e. value ends in
a 7 or an F), then it doesn't work. What seems to happen is that
the pha at the top of the print_str subroutine doesn't cause the
stack pointer to decrement in this case, although I am still trying
to verify that is what happens.

Strangely, if the stack pointer is set to $00, say, there is no
problem with the stack pointer passing through $FF, it just can't
start there!!??