Debugger program.
Has an NMI routine for sending information to LCD. Intended to be used
along with the BE6502 debugger board.

Routine initializes the LCD, clears the display and then prints out
the A, X, Y, and Status flags on the first line and the program
counter, the stack address and value on the top of the stack on the
second line. It then waits for a signal for IRQ from the debugger
board before returning from the NMI.

The main routine performs a variety of actions that test the
debugger board and NMI routine.